fix: delete unused class
