--- include/http/RequestValidator.hpp
@@ -1,27 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   RequestValidator.hpp                               :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: tsemenov <tsemenov@student.42berlin.de>    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/05/17 00:03:24 by tsemenov          #+#    #+#             */
-/*   Updated: 2026/05/20 21:34:00 by tsemenov         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#pragma once
-
-// #include <string>
-
-class HttpRequest;
-class ServerConfig;
-
-class RequestValidator {
-   private:
-    static int _checkRouteAndMethod(const HttpRequest& request, const ServerConfig& config);
-    static void _logResult(const HttpRequest& request, int code);
-
-   public:
-    static int validate(const HttpRequest& request, const ServerConfig& config);
-};
